refactor: checkList 디렉터리 명 수정
