+++ spring/src/main/java/com/life/jeonggiju/domain/checkListRecord/dto/FindCheckListResponse.java
@@ -1,4 +1,4 @@
-package com.life.jeonggiju.domain.checkList.dto;
+package com.life.jeonggiju.domain.checkListRecord.dto;
 
 import java.time.LocalDate;
 import java.util.UUID;
