+++ spring/src/main/java/com/life/jeonggiju/domain/checkListRecord/repository/CheckListRepository.java
@@ -1,4 +1,4 @@
-package com.life.jeonggiju.domain.checkList.repository;
+package com.life.jeonggiju.domain.checkListRecord.repository;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -7,7 +7,7 @@ import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.life.jeonggiju.domain.checkList.entity.CheckListRecord;
+import com.life.jeonggiju.domain.checkListRecord.entity.CheckListRecord;
 
 public interface CheckListRepository extends JpaRepository<CheckListRecord, UUID> {
